[WFCORE-7008] Split ManagementInterfaceResourcesTestCase for different stability levels.
