+++ testsuite/manualmode/pom.xml
@@ -31,9 +31,9 @@
         <!--
         <byteman.jvm.args>-javaagent:${org.jboss.byteman:byteman:jar}=port:${byteman.port},address:${byteman.host},boot:${org.jboss.byteman:byteman:jar} -Dorg.jboss.byteman.transform.all -Dorg.jboss.byteman.verbose=true</byteman.jvm.args>
         -->
-								
+
     </properties>
-    
+
     <dependencies>
         <dependency>
             <groupId>jakarta.inject</groupId>
@@ -583,7 +583,7 @@
                                         <!-- restart the server -->
                                         <exclude>org.jboss.as.test.manualmode.deployment.DeploymentScannerRedeploymentTestCase</exclude>
                                         <exclude>org.jboss.as.test.manualmode.deployment.DeploymentScannerUnitTestCase</exclude>
-                                        
+
                                         <!-- restart the server -->
                                         <exclude>org.jboss.as.test.manualmode.logging.LoggingDependenciesTestCase</exclude>
                                         <exclude>org.jboss.as.test.manualmode.logging.LoggingPreferencesTestCase</exclude>
@@ -598,7 +598,7 @@
                                         <exclude>org.jboss.as.test.manualmode.provisioning.InstallationManagerBootTestCase</exclude>
                                         <!-- restart the server -->
                                         <exclude>org.jboss.as.test.manualmode.management.*TestCase</exclude>
-                                        
+
                                         <exclude>org.jboss.as.test.manualmode.management.cli.ManagementOpTimeoutTestCase</exclude>
                                         <exclude>org.jboss.as.test.manualmode.management.cli.ShutdownTestCase</exclude>
 
@@ -615,21 +615,20 @@
                                         <exclude>org.jboss.as.test.manualmode.management.cli.RemoveManagementRealmTestCase</exclude>
                                         <!-- no git support -->
                                         <exclude>org.jboss.as.test.manualmode.management.persistence.*TestCase</exclude>
-                                        
+
                                         <!-- replace config file -->
                                         <exclude>org.jboss.as.test.manualmode.mgmt.elytron.ElytronModelControllerClientTestCase</exclude>
-                                        
+
                                         <!--- remove a module after having closed the server -->
                                         <exclude>org.jboss.as.test.manualmode.expressions.CredentialStoreExpressionsTestCase</exclude>
-                                        
+
                                         <!-- admin mode -->
-                                        <exclude>org.wildfly.core.test.standalone.mgmt.ManagementInterfaceResourcesTestCase</exclude>
                                         <exclude>org.jboss.as.test.manualmode.adminonly.auditlog.*TestCase.java</exclude>
                                         <exclude>org.jboss.as.test.manualmode.cli.boot.ops.CliBootOperationsTestCase.java</exclude>
 
                                         <!-- start /stop, and install extension before start -->
                                         <exclude>org.wildfly.core.test.standalone.mgmt.PreparedResponseTestCase</exclude>
-                                        
+
                                         <!-- start/stop/start the server -->
                                         <exclude>org.wildfly.core.test.standalone.mgmt.events.*TestCase</exclude>
                                         <!-- Requires modification of the process-uuid file inside the Jar -->
